+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_engine/routes/enable.ts
@@ -8,30 +8,28 @@
 import type { StartServicesAccessor } from '@kbn/core/server';
 import { buildSiemResponse } from '@kbn/lists-plugin/server/routes/utils';
 import { transformError } from '@kbn/securitysolution-es-utils';
-import { RISK_ENGINE_DISABLE_URL, APP_ID } from '../../../../common/constants';
-import type { StartPlugins } from '../../../plugin';
-import type { SecuritySolutionPluginRouter } from '../../../types';
+import { RISK_ENGINE_ENABLE_URL, APP_ID } from '../../../../../common/constants';
 import { TASK_MANAGER_UNAVAILABLE_ERROR } from './translations';
+import type { StartPlugins } from '../../../../plugin';
+import type { SecuritySolutionPluginRouter } from '../../../../types';
 
-export const riskEngineDisableRoute = (
+export const riskEngineEnableRoute = (
   router: SecuritySolutionPluginRouter,
   getStartServices: StartServicesAccessor<StartPlugins>
 ) => {
   router.versioned
     .post({
       access: 'internal',
-      path: RISK_ENGINE_DISABLE_URL,
+      path: RISK_ENGINE_ENABLE_URL,
       options: {
         tags: ['access:securitySolution', `access:${APP_ID}-entity-analytics`],
       },
     })
     .addVersion({ version: '1', validate: {} }, async (context, request, response) => {
       const siemResponse = buildSiemResponse(response);
-
       const [_, { taskManager }] = await getStartServices();
       const securitySolution = await context.securitySolution;
       const riskEngineClient = securitySolution.getRiskEngineDataClient();
-
       if (!taskManager) {
         return siemResponse.error({
           statusCode: 400,
@@ -40,7 +38,7 @@ export const riskEngineDisableRoute = (
       }
 
       try {
-        await riskEngineClient.disableRiskEngine({ taskManager });
+        await riskEngineClient.enableRiskEngine({ taskManager });
         return response.ok({ body: { success: true } });
       } catch (e) {
         const error = transformError(e);
