+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_engine/routes/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { riskEngineInitRoute } from './init';
+export { riskEngineEnableRoute } from './enable';
+export { riskEngineDisableRoute } from './disable';
+export { riskEngineStatusRoute } from './status';
