+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_engine/routes/init.ts
@@ -8,11 +8,11 @@
 import type { StartServicesAccessor, Logger } from '@kbn/core/server';
 import { buildSiemResponse } from '@kbn/lists-plugin/server/routes/utils';
 import { transformError } from '@kbn/securitysolution-es-utils';
-import { RISK_ENGINE_INIT_URL, APP_ID } from '../../../../common/constants';
-import type { StartPlugins } from '../../../plugin';
+import { RISK_ENGINE_INIT_URL, APP_ID } from '../../../../../common/constants';
+import type { StartPlugins } from '../../../../plugin';
 import { TASK_MANAGER_UNAVAILABLE_ERROR } from './translations';
-import type { SecuritySolutionPluginRouter } from '../../../types';
-import type { InitRiskEngineResultResponse } from '../types';
+import type { SecuritySolutionPluginRouter } from '../../../../types';
+import type { InitRiskEngineResultResponse } from '../../types';
 
 export const riskEngineInitRoute = (
   router: SecuritySolutionPluginRouter,
