+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_engine/routes/translations.ts
