+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/routes/calculation.test.ts
@@ -5,23 +5,23 @@
  * 2.0.
  */
 
-import { riskScoreCalculationRoute } from './risk_score_calculation_route';
+import { riskScoreCalculationRoute } from './calculation';
 
 import { loggerMock } from '@kbn/logging-mocks';
 
-import { RISK_SCORE_CALCULATION_URL } from '../../../../common/constants';
+import { RISK_SCORE_CALCULATION_URL } from '../../../../../common/constants';
 import {
   serverMock,
   requestContextMock,
   requestMock,
-} from '../../detection_engine/routes/__mocks__';
-import { riskScoreServiceFactory } from '../risk_score/risk_score_service';
-import { riskScoreServiceMock } from '../risk_score/risk_score_service.mock';
-import { getRiskInputsIndex } from '../risk_score/get_risk_inputs_index';
-import { calculateAndPersistRiskScoresMock } from '../risk_score/calculate_and_persist_risk_scores.mock';
+} from '../../../detection_engine/routes/__mocks__';
+import { riskScoreServiceFactory } from '../risk_score_service';
+import { riskScoreServiceMock } from '../risk_score_service.mock';
+import { getRiskInputsIndex } from '../get_risk_inputs_index';
+import { calculateAndPersistRiskScoresMock } from '../calculate_and_persist_risk_scores.mock';
 
-jest.mock('../risk_score/get_risk_inputs_index');
-jest.mock('../risk_score/risk_score_service');
+jest.mock('../get_risk_inputs_index');
+jest.mock('../risk_score_service');
 
 describe('risk score calculation route', () => {
   let server: ReturnType<typeof serverMock.create>;
