+++ x-pack/plugins/security_solution/server/lib/entity_analytics/risk_score/routes/preview.test.ts
@@ -7,20 +7,20 @@
 
 import { loggerMock } from '@kbn/logging-mocks';
 
-import { RISK_SCORE_PREVIEW_URL } from '../../../../common/constants';
-import { RiskCategories, RiskWeightTypes } from '../../../../common/risk_engine';
+import { RISK_SCORE_PREVIEW_URL } from '../../../../../common/constants';
+import { RiskCategories, RiskWeightTypes } from '../../../../../common/risk_engine';
 import {
   serverMock,
   requestContextMock,
   requestMock,
-} from '../../detection_engine/routes/__mocks__';
-import { getRiskInputsIndex } from '../risk_score/get_risk_inputs_index';
-import { riskScoreServiceFactory } from '../risk_score/risk_score_service';
-import { riskScoreServiceMock } from '../risk_score/risk_score_service.mock';
-import { riskScorePreviewRoute } from './risk_score_preview_route';
+} from '../../../detection_engine/routes/__mocks__';
+import { getRiskInputsIndex } from '../get_risk_inputs_index';
+import { riskScoreServiceFactory } from '../risk_score_service';
+import { riskScoreServiceMock } from '../risk_score_service.mock';
+import { riskScorePreviewRoute } from './preview';
 
-jest.mock('../risk_score/risk_score_service');
-jest.mock('../risk_score/get_risk_inputs_index');
+jest.mock('../risk_score_service');
+jest.mock('../get_risk_inputs_index');
 
 describe('POST risk_engine/preview route', () => {
   let server: ReturnType<typeof serverMock.create>;
