--- x-pack/plugins/security_solution/server/lib/entity_analytics/routes/index.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export { riskScorePreviewRoute } from './risk_score_preview_route';
-export { riskEngineInitRoute } from './risk_engine_init_route';
-export { riskEngineEnableRoute } from './risk_engine_enable_route';
-export { riskEngineDisableRoute } from './risk_engine_disable_route';
-export { riskEngineStatusRoute } from './risk_engine_status_route';
-export { riskEnginePrivilegesRoute } from './risk_engine_privileges_route';
